% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_editable_map.R
\name{empty_editable_map}
\alias{empty_editable_map}
\title{empty_editable_map}
\usage{
empty_editable_map(lon, lat, crs = 4326, sf_use_s2 = FALSE)
}
\arguments{
\item{lon}{Longitude limits.}

\item{lat}{Latitude limits.}

\item{crs}{Coordinate reference system (EPSG). By default the value is 4326, which corresponds to EPSG::4326 (WGS84)}

\item{sf_use_s2}{sf_use_s2 option, deafult is FALSE}
}
\value{
a map
}
\description{
Plot a world map which can be edited in powerpoint.
}
\examples{
\dontrun{empty_editable_map(lon = c(2600000, 9300000), lat = c(5450000, 1000000), crs = 3035)
export::graph2ppt(file ="editable_map.pptx", width = 7, height = 4)}
\dontrun{empty_editable_map(lon = c(-15, 40), lat = c(20, 65), crs = 4326)}

}
