% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries_data.R
\name{geometries_data}
\alias{geometries_data}
\title{geometries_data}
\usage{
geometries_data(exclude.iso.na = TRUE, countries.list = NULL)
}
\arguments{
\item{exclude.iso.na}{if TRUE (default), countries that do not have a iso code are excluded from the table}

\item{countries.list}{List of the iso_a2 codes of countries that are to be included. By default it is set to NULL and all countries are included}
}
\value{
a data frame
}
\description{
Generate a data frame with information about geometries and centroid coordinates of
all the countries or only a selected group.
}
\examples{
geometries_data(countries.list = c("IT", "FR", "US"))

}
