% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df0.R
\docType{data}
\name{map_df0}
\alias{map_df0}
\title{Initial dataset with geometries for each country}
\format{
An object of class \code{data.frame}
}
\usage{
data(map_df0)
}
\description{
This data set contains information about name, iso_a2 code and geometry for 242 countries.
}
\examples{
data(map_df0)
head(map_df0)
}
\keyword{datasets}
