% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_ITBIH.R
\docType{data}
\name{regions_ITBIH}
\alias{regions_ITBIH}
\title{Some potentially useful regions for more granular maps}
\format{
An object of class \code{data.frame}
}
\usage{
data(regions_ITBIH)
}
\description{
This data set contains information about name, iso_a2 code and geometry for Italian and BIH regions.
}
\examples{
data(regions_ITBIH)
head(regions_ITBIH)
}
\keyword{datasets}
