% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata1b.R
\docType{data}
\name{testdata1b}
\alias{testdata1b}
\title{Simulated data set 1b}
\format{
An object of class \code{data.frame}
}
\usage{
data(testdata1b)
}
\description{
Data from a random simulation with continuous and categorical data.
}
\examples{
data(testdata1b)
head(testdata1b)
}
\keyword{datasets}
