% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata1c.R
\docType{data}
\name{testdata1c}
\alias{testdata1c}
\title{Simulated data set 1c}
\format{
An object of class \code{data.frame}
}
\usage{
data(testdata1c)
}
\description{
Data from a random simulation with continuous and categorical data. This data set contains information
about 237 countries (countries without unique ISO 3166 code are excluded).
}
\examples{
data(testdata1c)
head(testdata1c)
}
\keyword{datasets}
