% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldMapR-package.R
\docType{package}
\name{WorldMapR-package}
\alias{WorldMapR}
\alias{WorldMapR-package}
\title{WorldMapR: Worldwide or Coordinates-Based Heat Maps}
\description{
Easily plot heat maps of the world, based on continuous or categorical data. Country labels can also be added to the map.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Luigi-Annic/WorldMapR/}
  \item Report bugs at \url{https://github.com/Luigi-Annic/WorldMapR/issues}
}

}
\author{
\strong{Maintainer}: Luigi Annicchiarico \email{luigi.annic@gmail.com}

}
\keyword{internal}
