% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldplot.R
\name{worldplot}
\alias{worldplot}
\title{worldplot}
\usage{
worldplot(
  data,
  div = 1,
  ColName,
  CountryName,
  CountryNameType = "isoa2",
  rangeVal,
  longitude = c(-180, 180),
  latitude = c(-90, 90),
  title = "",
  legendTitle = as.character(ColName),
  annote = FALSE
)
}
\arguments{
\item{data}{Data set containing the list of nations and the variable that we want to plot}

\item{div}{Controlling image quality (and image size). Default value is 1}

\item{ColName}{character variable with the name of the variable of interest}

\item{CountryName}{character variable with the name of the country names column}

\item{CountryNameType}{character variable with the coding for CountryName. It can be "isoa2" (default), "isoa3", or "name"}

\item{rangeVal}{limit values that are to be defined for the map}

\item{longitude}{longitude limits. Default is c(-180, 180) (whole world)}

\item{latitude}{latitude limits. Default is c(-90, 90) (whole world)}

\item{title}{title of the plot. Default is no title}

\item{legendTitle}{title of the legend. Default is the name of the filling variable}

\item{annote}{do you want to plot country labels (iso2 code) on the map?}
}
\value{
a map
}
\description{
Plot a world map for continuous data
}
\examples{
data(testdata1b)
worldplot(data = testdata1b,
          div = 1,
          ColName = "VNum",
          CountryName = "Cshort",
          CountryNameType = "isoa2",
          rangeVal = c(0,50),
          annote = FALSE)

}
