% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldplotCat.R
\name{worldplotCat}
\alias{worldplotCat}
\title{worldplotCat}
\usage{
worldplotCat(
  data,
  ColName,
  CountryName,
  CountryNameType,
  longitude = c(-180, 180),
  latitude = c(-90, 90),
  title = "",
  legendTitle = as.character(ColName),
  Categories = levels(factor(map_df$MapFiller)),
  annote = FALSE,
  div = 1
)
}
\arguments{
\item{data}{Data set containing the list of nations and the variable that we want to plot}

\item{ColName}{character variable with the name of the variable of interest}

\item{CountryName}{character variable with the name of the country names column}

\item{CountryNameType}{character variable with the coding for `CountryName`. It can be "isoa2" (default), "isoa3", or "name"}

\item{longitude}{longitude limits. Default is `c(-180, 180)` (whole world)}

\item{latitude}{latitude limits. Default is `c(-90, 90)` (whole world)}

\item{title}{title of the plot. Default is no title}

\item{legendTitle}{title of the legend. Default is the name of the filling variable}

\item{Categories}{categories labels to be plotted in the legend}

\item{annote}{do you want to plot country labels (iso2 code) on the map? Default is set to `FALSE`}

\item{div}{Controlling image quality (and image size). Default value is 1}
}
\value{
a map
}
\description{
Plot a world map for categorical data
}
\examples{
data(testdata1b)
worldplotCat(data = testdata1b,
             div = 1,
             ColName = "VCat",
             CountryName = "Cshort",
             CountryNameType = "isoa2",
             annote = FALSE)

}
