% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries_data.R
\name{geometries_data}
\alias{geometries_data}
\title{geometries_data}
\usage{
geometries_data(exclude.iso.na = TRUE, countries.list = NULL)
}
\arguments{
\item{exclude.iso.na}{if `TRUE` (default), countries that do not have a iso code are excluded from the table}

\item{countries.list}{List of the iso a2 codes of countries that are to be included. By default it is set to `NULL` and all countries are included}
}
\value{
an object of class `data.frame` and `sf`
}
\description{
This function generates a data frame with information about geometries and centroid coordinates of
countries. You can choose whether to keep all the countries or a subset.
}
\examples{
geometries_data(countries.list = c("IT", "FR", "US"))

}
