% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata1.R
\docType{data}
\name{testdata1}
\alias{testdata1}
\title{Simulated data set 1}
\format{
An object of class \code{data.frame}
}
\usage{
data(testdata1)
}
\description{
Data from a random simulation with continuous data.
}
\examples{
data(testdata1)
head(testdata1)
}
\keyword{datasets}
