% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldplot.R
\name{worldplot}
\alias{worldplot}
\title{worldplot}
\usage{
worldplot(
  data,
  ColName,
  CountryName,
  CountryNameType = "isoa2",
  rangeVal,
  longitude = c(-180, 180),
  latitude = c(-90, 90),
  crs = NULL,
  title = "",
  legendTitle = as.character(ColName),
  annote = FALSE,
  div = 1,
  palette_option = "D",
  save = FALSE,
  filename = "worldplot.jpg",
  path = tempdir(),
  width = 20,
  height = 10,
  units = "cm",
  scale = 1
)
}
\arguments{
\item{data}{Data set containing the list of nations and the variable that we want to plot.}

\item{ColName}{Character variable with the name of the variable of interest.}

\item{CountryName}{Character variable with the name of the country names column.}

\item{CountryNameType}{Character variable with the coding for \code{CountryName}. One of \code{isoa2} (default), \code{isoa3}, or \code{name}.}

\item{rangeVal}{Limit values that are to be defined for the map.}

\item{longitude}{Longitude limits. Default is \code{c(-180, 180)} (whole world with crs = 4326).}

\item{latitude}{Latitude limits. Default is \code{c(-90, 90)} (whole world with crs = 4326).}

\item{crs}{Coordinate reference system. By default the value is equal to NULL, which corresponds to crs = 4326}

\item{title}{Title of the plot. Default is no title.}

\item{legendTitle}{Title of the legend. Default is the name of the filling variable.}

\item{annote}{Do you want to plot country labels (ISO 3166-1 alpha-2 code) on the map? Default is set to \code{FALSE}.}

\item{div}{Parameter for modifying the elements dimensions in the map. Usually, it does not need to be modified. Default value is 1.}

\item{palette_option}{Character string indicating the palette to be used. Available options range between "A" and "H".}

\item{save}{Removed since v 0.1.3. Save the plot in a jpg file?}

\item{filename}{Removed since v 0.1.3. Only if is save set to \code{TRUE}. Name of the file.}

\item{path}{Removed since v 0.1.3. Only if save is set to \code{TRUE}. Path of the directory where the file is to be saved.}

\item{width}{Removed since v 0.1.3. Only if save is set to \code{TRUE}. Width of the file.}

\item{height}{Removed since v 0.1.3. Only if save is set to \code{TRUE}. Height of the file.}

\item{units}{Removed since v 0.1.3. Only if save is set to \code{TRUE}. Units for width and height. Can be 'cm', 'mm', 'in', or 'px'.}

\item{scale}{Removed since v 0.1.3. Only if save is set to \code{TRUE}. Scaling factor for adjusting image dimensions.}
}
\value{
a map
}
\description{
Plot a world heat map based on a continuous variable.
}
\examples{
data(testdata1b)
worldplot(data = testdata1b,
          div = 1,
          ColName = "VNum",
          CountryName = "Cshort",
          CountryNameType = "isoa2",
          rangeVal = c(0,50),
          annote = FALSE)

}
